requirements.txt
kafka-python==2.0.2
Flask==2.2.2
Werkzeug==2.2.2
pyspark==3.3.0
SQLAlchemy==1.4.47
psycopg2-binary==2.9.3
apache-airflow==2.6.1
pandas==1.5.3
python-dotenv==0.19.1
